Move nodemailer.js to email
